+++ app/Http/Controllers/Api/PersonController.php
@@ -2,17 +2,15 @@
 
 namespace App\Http\Controllers\Api;
 
+
 use App\Contracts\Finders\FinderCollection;
 use App\Http\Requests\Api\PersonStoreRequest;
 use App\Http\Requests\Api\PersonUpdateRequest;
 use App\Http\Resources\Api\PersonResource;
-use App\Http\Resources\Api\Resource;
 use App\Person;
 use App\Services\Finders\GroupFinder;
 use App\Services\Sorters\PersonSorter;
 use Illuminate\Http\Request;
-use Illuminate\Http\Resources\Json\ResourceCollection;
-
 class PersonController extends Controller
 {
 
@@ -43,9 +41,12 @@ class PersonController extends Controller
      * @param  \App\Person  $person
      * @param  Request      $request
      * @return Resource
+     * @throws
      */
     public function show(Person $person, Request $request)
     {
+        $this->authorize('view', $person);
+
         return $this->prepare($person, $request);
     }
 
@@ -73,6 +74,8 @@ class PersonController extends Controller
      */
     public function destroy(Person $person)
     {
+        $this->authorize('delete', $person);
+
         $person->delete();
     }
