+++ app/Http/Controllers/Api/UserController.php
@@ -11,19 +11,9 @@ use Illuminate\Validation\Rule;
 
 class UserController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @param Request $request
-     * @return ResourceCollection
-     */
-    public function index(Request $request)
-    {
-        $query = User::query();
-        $query->with($this->getAskedRelations($request));
 
-        return UserResource::collection($query->paginate());
-    }
+    protected $modelClass = User::class;
+    protected $resourceClass = UserResource::class;
 
     /**
      * Store a new User in the database.
@@ -35,6 +25,8 @@ class UserController extends Controller
      */
     public function store(Request $request, FinderCollection $finders)
     {
+        $this->authorize('store', User::class);
+
         $validatedData = $request->validate([
             'name' => 'required|unique:users|string|max:255',
             'email' => 'required|unique:users|email|max:255',
@@ -65,9 +57,12 @@ class UserController extends Controller
      * @param  Request    $request
      * @param  \App\User  $user
      * @return UserResource
+     * @throws
      */
     public function show(Request $request, User $user)
     {
+        $this->authorize('view', $user);
+
         $user->load($this->getAskedRelations($request));
         return new UserResource($user);
     }
@@ -83,6 +78,8 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user, FinderCollection $finders)
     {
+        $this->authorize('update', $user);
+
         $validatedData = $request->validate([
             'name' => ['sometimes','required','string','max:255', Rule::unique('users')->ignore($user->id)],
             'email' => ['sometimes','required','email','max:255', Rule::unique('users')->ignore($user->id)],
@@ -120,10 +117,12 @@ class UserController extends Controller
 
     /**
      * @param User $user
-     * @throws \Exception
+     * @throws
      */
     public function destroy(User $user)
     {
+        $this->authorize('delete', $user);
+
         $user->delete();
     }
 }
